requirements.txt
python-telegram-bot==13.7
python-dotenv==0.18.0
gunicorn==20.1.0